requirements.txt
matplotlib
streamlit
python-docx